Remove function pointer for DOSBOX_Restart
