+++ src/dosbox.cpp
@@ -11,6 +11,10 @@
 #include <limits>
 #include <memory>
 
+#ifdef WIN32
+#include <windows.h>
+#endif
+
 #include "audio/disk_noise.h"
 #include "audio/mixer.h"
 #include "capture/capture.h"
@@ -513,21 +517,140 @@ void DOSBOX_SetMachineTypeFromConfig(SectionProp& section)
 	       (machine != MachineType::Vga && svga_type == SvgaType::None));
 }
 
-static DosboxRestartFunction restart_func = {};
-
-void DOSBOX_SetRestartFunction(DosboxRestartFunction func)
+static void remove_waitpid(std::vector<std::string>& parameters)
 {
-	restart_func = func;
+	auto it = parameters.begin();
+	while (it != parameters.end()) {
+		if (*it == "-waitpid" || *it == "--waitpid") {
+			it = parameters.erase(it);
+			if (it != parameters.end()) {
+				auto& pid = *it;
+				if (!pid.empty() && std::isdigit(pid[0])) {
+					// The integer value should be the next
+					// element following "--waitpid". If we
+					// found an integer, remove it as well.
+					it = parameters.erase(it);
+				}
+			}
+		} else {
+			++it;
+		}
+	}
 }
 
 void DOSBOX_Restart()
 {
-	restart_func(control->startup_params);
+	DOSBOX_Restart(control->startup_params);
 }
 
 void DOSBOX_Restart(std::vector<std::string>& parameters)
 {
-	restart_func(parameters);
+	control->ApplyQueuedValuesToCli(parameters);
+
+	// Remove any existing --waitpid parameters.
+	// This can happen with multiple restarts.
+	remove_waitpid(parameters);
+
+#ifdef WIN32
+	parameters.emplace_back("--waitpid");
+	parameters.emplace_back(std::to_string(GetCurrentProcessId()));
+	std::string command_line = {};
+
+	bool first = true;
+
+	for (const auto& arg : parameters) {
+		if (!first) {
+			command_line.push_back(' ');
+		}
+		command_line.append(arg);
+		first = false;
+	}
+#else
+	parameters.emplace_back("--waitpid");
+	parameters.emplace_back(std::to_string(getpid()));
+
+	char** newargs = new char*[parameters.size() + 1];
+
+	// parameter 0 is the executable path
+	// contents of the vector follow
+	// last one is NULL
+	for (size_t i = 0; i < parameters.size(); i++) {
+		newargs[i] = parameters[i].data();
+	}
+	newargs[parameters.size()] = nullptr;
+#endif // WIN32
+
+	GFX_RequestExit(true);
+
+#if C_DEBUGGER
+	// shutdown curses
+	DEBUG_Destroy();
+#endif
+
+#ifdef WIN32
+	// nullptr to parse from command line
+	const LPCSTR application_name = nullptr;
+
+	// nullptr for default
+	const LPSECURITY_ATTRIBUTES process_attributes = nullptr;
+
+	// nullptr for default
+	const LPSECURITY_ATTRIBUTES thread_attributes = nullptr;
+
+	const BOOL inherit_handles = FALSE;
+
+	// CREATE_NEW_CONSOLE fixes a bug where the parent process exiting kills
+	// the child process. This can manifest when you use the "restart"
+	// hotkey action to restart DOSBox.
+	// https://github.com/dosbox-staging/dosbox-staging/issues/3346
+	const DWORD creation_flags = CREATE_NEW_CONSOLE;
+
+	// nullptr to use parent's environment
+	const LPVOID environment_variables = nullptr;
+
+	// nullptr to use parent's current directory
+	const LPCSTR current_directory = nullptr;
+
+	// Input structure with a bunch of stuff we probably don't care about.
+	// Zero it out to use defaults save for the size field.
+	STARTUPINFO startup_info = {};
+	startup_info.cb          = sizeof(startup_info);
+
+	// Output structure we also don't care about.
+	PROCESS_INFORMATION process_information = {};
+
+	if (!CreateProcess(application_name,
+	                   const_cast<char*>(command_line.c_str()),
+	                   process_attributes,
+	                   thread_attributes,
+	                   inherit_handles,
+	                   creation_flags,
+	                   environment_variables,
+	                   current_directory,
+	                   &startup_info,
+	                   &process_information)) {
+		LOG_ERR("Restart failed: CreateProcess failed");
+	}
+#else
+	int ret = fork();
+	switch (ret) {
+	case -1:
+		LOG_ERR("Restart failed: fork failed: %s", strerror(errno));
+		break;
+	case 0:
+		// Newly created child process immediately executes a new
+		// instance of DOSBox. It is not safe for a child process in a
+		// multi-threaded program to do much else.
+		if (execvp(newargs[0], newargs) == -1) {
+			E_Exit("Restart failed: execvp failed: %s", strerror(errno));
+		}
+		break;
+	}
+	// Original (parent) process continues execution here.
+	// It will proceed to do a normal shutdown due to
+	// gfx_request_exit(true); above.
+	delete[] newargs;
+#endif // WIN32
 }
 
 static void dosbox_realinit(SectionProp& section)
